# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "guru-core API",
    "description": "\nThe single endpoint the app talks to. A user states a goal; the service collects\nwhatever context exists, asks follow-up questions when something essential is missing,\nand produces three difficulty variants of a scheduled plan.\n\n**Authentication.** Every endpoint except `POST /v1/auth/google`, `GET /health` and the\npresigned `/v1/files/*` routes needs `Authorization: Bearer <jwt>`. Get the token from\n`POST /v1/auth/google`; it carries the user id and expires after `JWT_TTL_SECONDS`.\nThe team-facing `role-models` writes authenticate with `X-API-Key` instead.\n\n**Long-running work.** Plan generation, import parsing and calendar export run on a\nqueue, so the endpoints that start them return `202` with an id. Poll the resource\n(`GET /v1/plan-sessions/{id}`, `GET /v1/imports`) rather than expecting a synchronous\nresult — PostgreSQL is the source of truth for job state, Redis only caches it.\n\n**Errors.** Every failure returns the same envelope:\n`{\"error\": {\"code\": \"not_found\", \"message\": \"...\"}}`. The code is the snake_case name of\nthe domain error: `invalid_input` (422), `unauthorized` (401), `forbidden` (403),\n`not_found` (404), `conflict` (409), `reauth_required` (409), `rate_limited` (429).\n",
    "contact": {
      "name": "Quasar-Gang",
      "url": "https://github.com/Quasar-Gang/guru-core"
    },
    "license": {
      "name": "Proprietary — all rights reserved"
...
  "paths": {
...
    "/v1/imports/google-calendar": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import upcoming Google Calendar events as context",
        "description": "Pull the user's existing commitments straight from Google, with no file involved.\n\n**Prerequisite:** the user must already have connected Google through\n`GET /v1/integrations/google/authorize` and `POST /v1/integrations/google/callback`.\nSigning in with Google is not enough — calendar access is a separate consent. Without\nit this returns `409 reauth_required`.\n\nUnlike an upload this is **synchronous**: the events are fetched and stored within the\nrequest, and the returned import is already `parsed` with `event_count` filled in.\nThere is nothing to poll and no presign/complete pair.\n\n`days` is a forward-looking window starting now — default `90`, maximum `365`. Past\nevents are never imported. Re-running creates a new, independent import rather than\nupdating the previous one, so plan generation sees the newest pull alongside the old.",
        "operationId": "import_google_calendar_v1_imports_google_calendar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleCalendarImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The finished import, already in status `parsed`, with its event count.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — missing, invalid or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "`reauth_required` — Google is not connected for this user, the connection was disconnected, or the stored refresh token was rejected. Send the user through `GET /v1/integrations/google/authorize` again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — `days` is outside the range 1–365.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ]
      }
    },
    "/v1/integrations": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "List the user's third-party connections",
        "description": "What the app checks before offering anything that needs Google.\n\nRead `connected` and `needs_reauth` together:\n\n- not listed at all — never connected; show a \"Connect Google\" button.\n- `connected: true` — good to go for calendar import and calendar/Sheets export.\n- `connected: false, needs_reauth: true` — the connection was disconnected or revoked\n  upstream. The row is deliberately kept so the app can prompt for a reconnect; the\n  fix is the same authorize → callback pair as a first connection.\n\n`scopes` lists what Google actually granted, and `connected_at` is when the row was\nfirst created. An empty list is normal: signing in with Google creates no integration\nrow, because sign-in and calendar access are separate grants.",
        "operationId": "list_integrations_v1_integrations_get",
        "responses": {
          "200": {
            "description": "One row per provider the user has ever connected.",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IntegrationView"
                  },
                  "type": "array",
                  "title": "Response List Integrations V1 Integrations Get"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — missing, invalid or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ]
      }
    },
    "/v1/integrations/{provider}/authorize": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Step 1 of connecting a provider: get the consent URL",
        "description": "Begin the second Google grant — the one that gives the backend calendar access.\n\n`provider` must be `google`; it is the only integration in the MVP.\n\nSend the user to `authorize_url` (redirect or popup). One consent covers all three\nthings the backend does on their behalf: reading the calendar for\n`POST /v1/imports/google-calendar`, writing events for calendar export, and creating\nthe spreadsheet for the Sheets export. It is requested with `access_type=offline` and\n`prompt=consent` so Google returns a refresh token; without one the connection cannot\nbe completed.\n\nThe URL already carries an unguessable `state` nonce and the server-configured\nredirect URI — pass it through unmodified. Nothing is persisted until step 2,\n`POST /v1/integrations/{provider}/callback`, so an abandoned consent leaves no trace.",
        "operationId": "authorize_integration_v1_integrations__provider__authorize_get",
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A Google consent URL to send the user to. Nothing is stored yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeUrlResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — missing, invalid or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — `provider` is not `google`, the only provider supported.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/integrations/{provider}/callback": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Step 2 of connecting a provider: redeem the consent code",
        "description": "Finish the connection with the `code` Google appended to the redirect.\n\nNote this is **the app's own authenticated endpoint**, not a URL Google calls: the\nclient catches the redirect, pulls `code` out of the query string, and posts it here\nwith its normal bearer token. That is why the `state` nonce only has to be\nunguessable — the user is identified by the JWT, not by `state`.\n\nThe refresh token is encrypted at rest and never leaves the backend; the client is\nnever given a Google token. A repeat connection upserts the same row rather than\ncreating a duplicate, so reconnecting after a `needs_reauth` is safe.\n\nOnce this returns, `POST /v1/imports/google-calendar` and the Google export routes\nstop failing with `409 reauth_required`.",
        "operationId": "complete_integration_v1_integrations__provider__callback_post",
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The stored connection, now `connected: true`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — missing, invalid or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — `provider` is not `google`, or Google returned no refresh token (usually a re-consent where the user had already granted access). Restart at `/authorize`, which forces the consent screen.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "The token exchange with Google failed outright — typically a `code` that is expired, already redeemed, or issued for a different redirect URI. Restart at `/authorize` for a fresh code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/integrations/{provider}": {
      "delete": {
        "tags": [
          "integrations"
        ],
        "summary": "Disconnect a provider",
        "description": "Revoke the grant upstream at Google and stop using it here.\n\nThe stored refresh token is revoked with Google, the row is marked revoked, and the\ncached access token is dropped, so the very next Google call fails with\n`409 reauth_required` rather than using a stale token.\n\nThe row is **not** deleted: `GET /v1/integrations` keeps returning it as\n`connected: false, needs_reauth: true` so the app can offer a reconnect. Calling this\ntwice is safe — the second call finds the row already revoked and only re-clears the\ncache. Plans already exported to Google Calendar are left in place; only future\nimports and exports stop working.",
        "operationId": "disconnect_integration_v1_integrations__provider__delete",
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Disconnected. Nothing is returned."
          },
          "401": {
            "description": "`unauthorized` — missing, invalid or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — this user has no connection row for that provider.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — `provider` is not `google`, the only provider supported.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorizeUrlResponse": {
        "properties": {
          "authorize_url": {
            "type": "string",
            "title": "Authorize Url"
          }
        },
        "type": "object",
        "required": [
          "authorize_url"
        ],
        "title": "AuthorizeUrlResponse"
      },
      "CallbackRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "CallbackRequest"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "GoogleCalendarImportRequest": {
        "properties": {
          "days": {
            "type": "integer",
            "title": "Days",
            "default": 90
          }
        },
        "type": "object",
        "title": "GoogleCalendarImportRequest"
      },
...
      "ImportView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count",
            "default": 0
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "format",
          "filename",
          "status",
          "error",
          "created_at"
        ],
        "title": "ImportView",
        "description": "One import as seen by the API client."
      },
      "IntegrationView": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "needs_reauth": {
            "type": "boolean",
            "title": "Needs Reauth"
          },
          "connected_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connected At"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "connected",
          "scopes",
          "needs_reauth",
          "connected_at"
        ],
        "title": "IntegrationView",
        "description": "One provider connection as seen by the API client (PRD 3.6)."
      },
...
    },
    "securitySchemes": {
      "BearerJWT": {
        "type": "http",
        "description": "A JWT issued by `POST /v1/auth/google`. Send it as `Authorization: Bearer <token>`.",
        "scheme": "bearer"
      }
    }
  },
...
  "servers": [
...
}